[libromdata] Wim: Should cast size to int, not size_t.
LibRpText::utf16le_to_utf8() takes an int length.

Fixes a 64-bit MSVC warning:

src\libromdata\Other\Wim.cpp(203,82): warning C4267: 'argument':
conversion from 'size_t' to 'int', possible loss of data

+++ src/libromdata/Other/Wim.cpp
@@ -200,7 +200,7 @@ int WimPrivate::addFields_XML()
 
 	// the xml inside wims are utf-16 but tinyxml only supports utf-8
 	// this means we have to do some conversion
-	string utf8_xml = LibRpText::utf16le_to_utf8(xml_data, static_cast<size_t>(size / 2));
+	string utf8_xml = LibRpText::utf16le_to_utf8(xml_data, static_cast<int>(size / 2));
 	delete[] xml_data;
 
 	XMLDocument document;
